refactor: createArticleSection and read will no longer create if does not exist
